pyproject.toml
[tool.poetry]
name = "ilpas"
version = "0.1.0"
description = ""
authors = ["Maguire <papaymaguire@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
hatchet-sdk = "^0.47.0"
click = {version = "^8.1.8", optional = true}
httpx = "^0.28.1"
cryptography = "^44.0.2"

[tool.poetry.extras]
cli=["click"]

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
testcontainers = "^4.9.0"
mypy = "^1.13.0"
uvicorn = "^0.34.0"
pytest-cov = "^6.0.0"

[tool.pytest.ini_options]
# Basic pytest configuration
testpaths = ["tests"]
python_files = "*_test.py"
# For pytest-asyncio
asyncio_mode = "auto"

[tool.coverage.run]
source = ["ilpas"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]

# Optional HTML report settings
[tool.coverage.html]
directory = "coverage_html_report"

[tool.pyright]
strict = true
reportMissingTypeStubs = true
reportUnboundVariable = true
reportMissingImports = false
reportOptionalMemberAccess = true
reportGeneralTypeIssues = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
